prepare add new collection
