+++ client/mock/ecommerce.edn
@@ -1,4 +1,20 @@
-{:slideshow
+{:userCart
+ [{:id  "cart_1557718717372_07617742",
+   :user_id  "ioobot",
+   :product_id  "1",
+   :specificationStock_id  "3",
+   :count 1,
+   :createdAt  "2019-05-13 11:38:37",
+   :updatedAt  ""
+   }
+  {:id "cart_1557718728636_91401876",
+   :user_id "ioobot",
+   :product_id "2",
+   :specificationStock_id "4",
+   :count 1,
+   :createdAt "2019-05-13 11:38:48",
+   :updatedAt ""}]
+ :slideshow
  [{:id "1",
   :name "1",
   :img
