+++ client/src/services/api.js
@@ -167,10 +167,10 @@ export async function deleteUser (params) {
   return QL.remove('user', params, ['result']);
 }
 export async function querySpecificationstock (params) {
-  return QL.find_many('specificationStock', deleteEmptyProperty(params), [  'stock',  'detailImg',  'id',  'status',  'slideImg',  'size',  'createdAt',  'updatedAt',  'color',  'orderproduct.orderPay_id',  'orderproduct.productPay',  'orderproduct.count',  'orderproduct.id',  'orderproduct.productPrice',  'orderproduct.updatedAt',  'orderproduct.productColor',  'orderproduct.unit',  'orderproduct.productSize',  'orderproduct.orderPay',  'orderproduct.createdAt',  'orderproduct.productImg',  'orderproduct.productName',  'orderproduct.specificationStock_id.updatedAt',  'orderproduct.specificationStock_id.color',  'orderproduct.specificationStock_id.createdAt',  'orderproduct.specificationStock_id.size',  'orderproduct.specificationStock_id.slideImg',  'orderproduct.specificationStock_id.status',  'orderproduct.specificationStock_id.id',  'orderproduct.specificationStock_id.detailImg',  'orderproduct.specificationStock_id.stock',  'orderproduct.product.recommend',  'orderproduct.product.updatedAt',  'orderproduct.product.unit',  'orderproduct.product.name',  'orderproduct.product.createdAt',  'orderproduct.product.status',  'orderproduct.product.id',  'orderproduct.product.intro',  'orderproduct.product.discountRate',  'orderproduct.product.price',  'orderproduct.product.img',  'orderproduct.product.stock',  'orderproduct.order.remark',  'orderproduct.order.updatedAt',  'orderproduct.order.orderLogistics_id',  'orderproduct.order.orderTotalPay',  'orderproduct.order.createdAt',  'orderproduct.order.orderStatus',  'orderproduct.order.id',  'orderproduct.order.count',  'orderproduct.order.productTotalPay',  'orderproduct.order.orderPay_id',  'orderproduct.user.email',  'orderproduct.user.updatedAt',  'orderproduct.user.password',  'orderproduct.user.telephone',  'orderproduct.user.username',  'orderproduct.user.createdAt',  'orderproduct.user.id',  'orderproduct.user.userData_id',  'product.stock',  'product.img',  'product.price',  'product.discountRate',  'product.intro',  'product.id',  'product.status',  'product.createdAt',  'product.name',  'product.unit',  'product.updatedAt',  'product.recommend',  'product.category.updatedAt',  'product.category.createdAt',  'product.category.status',  'product.category.order',  'product.category.img',  'product.category.id',  'product.category.name',  'product.orderproduct.updatedAt',  'product.orderproduct.productColor',  'product.orderproduct.unit',  'product.orderproduct.productSize',  'product.orderproduct.orderPay',  'product.orderproduct.createdAt',  'product.orderproduct.productImg',  'product.orderproduct.productName',  'product.orderproduct.productPrice',  'product.orderproduct.id',  'product.orderproduct.count',  'product.orderproduct.productPay',  'product.orderproduct.orderPay_id',  'product.specificationstock.updatedAt',  'product.specificationstock.color',  'product.specificationstock.createdAt',  'product.specificationstock.size',  'product.specificationstock.slideImg',  'product.specificationstock.status',  'product.specificationstock.id',  'product.specificationstock.detailImg',  'product.specificationstock.stock',  ]);
+  return QL.find_many('specificationStock', deleteEmptyProperty(params), [  'stock',  'detailImg',  'id',  'status',  'slideImg',  'size',  'createdAt',  'updatedAt',  'color']);
 }
 export async function queryOneSpecificationstock (params) {
-  return QL.find_one('specificationStock', params, [  'stock',  'detailImg',  'id',  'status',  'slideImg',  'size',  'createdAt',  'updatedAt',  'color',  'orderproduct.orderPay_id',  'orderproduct.productPay',  'orderproduct.count',  'orderproduct.id',  'orderproduct.productPrice',  'orderproduct.updatedAt',  'orderproduct.productColor',  'orderproduct.unit',  'orderproduct.productSize',  'orderproduct.orderPay',  'orderproduct.createdAt',  'orderproduct.productImg',  'orderproduct.productName',  'orderproduct.specificationStock_id.updatedAt',  'orderproduct.specificationStock_id.color',  'orderproduct.specificationStock_id.createdAt',  'orderproduct.specificationStock_id.size',  'orderproduct.specificationStock_id.slideImg',  'orderproduct.specificationStock_id.status',  'orderproduct.specificationStock_id.id',  'orderproduct.specificationStock_id.detailImg',  'orderproduct.specificationStock_id.stock',  'orderproduct.product.recommend',  'orderproduct.product.updatedAt',  'orderproduct.product.unit',  'orderproduct.product.name',  'orderproduct.product.createdAt',  'orderproduct.product.status',  'orderproduct.product.id',  'orderproduct.product.intro',  'orderproduct.product.discountRate',  'orderproduct.product.price',  'orderproduct.product.img',  'orderproduct.product.stock',  'orderproduct.order.remark',  'orderproduct.order.updatedAt',  'orderproduct.order.orderLogistics_id',  'orderproduct.order.orderTotalPay',  'orderproduct.order.createdAt',  'orderproduct.order.orderStatus',  'orderproduct.order.id',  'orderproduct.order.count',  'orderproduct.order.productTotalPay',  'orderproduct.order.orderPay_id',  'orderproduct.user.email',  'orderproduct.user.updatedAt',  'orderproduct.user.password',  'orderproduct.user.telephone',  'orderproduct.user.username',  'orderproduct.user.createdAt',  'orderproduct.user.id',  'orderproduct.user.userData_id',  'product.stock',  'product.img',  'product.price',  'product.discountRate',  'product.intro',  'product.id',  'product.status',  'product.createdAt',  'product.name',  'product.unit',  'product.updatedAt',  'product.recommend',  'product.category.updatedAt',  'product.category.createdAt',  'product.category.status',  'product.category.order',  'product.category.img',  'product.category.id',  'product.category.name',  'product.orderproduct.updatedAt',  'product.orderproduct.productColor',  'product.orderproduct.unit',  'product.orderproduct.productSize',  'product.orderproduct.orderPay',  'product.orderproduct.createdAt',  'product.orderproduct.productImg',  'product.orderproduct.productName',  'product.orderproduct.productPrice',  'product.orderproduct.id',  'product.orderproduct.count',  'product.orderproduct.productPay',  'product.orderproduct.orderPay_id',  'product.specificationstock.updatedAt',  'product.specificationstock.color',  'product.specificationstock.createdAt',  'product.specificationstock.size',  'product.specificationstock.slideImg',  'product.specificationstock.status',  'product.specificationstock.id',  'product.specificationstock.detailImg',  'product.specificationstock.stock',  ]);
+  return QL.find_one('specificationStock', params, [  'stock',  'detailImg',  'id',  'status',  'slideImg',  'size',  'createdAt',  'updatedAt',  'color'  ]);
 }
 export async function createSpecificationstock (params) {
   return QL.insert('specificationStock', deleteEmptyProperty(params), ['result']);
